backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
sse-starlette==2.0.0
PyJWT==2.8.0
pymongo==4.8.0
